==> Sentimental analysis Project/2. Pre-Processing.ipynb <==
# Mount Drive

from google.colab import drive
drive.mount('/content/drive')

# import libraries

import pandas as pd
import numpy as np
import re
import nltk
nltk.download('all')
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer

# Read data

data = pd.read_csv('/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/Data/Amazon_Customer_Reviews.csv')
data.drop('Unnamed: 0',axis=1,inplace=True)
data.head()

# Ratings unique value count

data['Ratings'].value_counts()

# Initialize word lemmatizer

wordnet = WordNetLemmatizer()

# Function for preprocessing

def preprocessing(data):    
    
    reviews =  []
    for i in range(0, len(data)):
        
        # Replacing values in rows
        review = data['Reviews'][i]
        review = review.replace('$','dollars')
    
        # keeping only text and numbers in reviews
        review = re.sub('%', ' percent', review)
        review = re.sub('[^a-zA-Z0-9/]', ' ', review)
        review = review.lower()
        review = review.split()
    
        # Removing Stopwords
        review = [wordnet.lemmatize(word) for word in review if not word in stopwords.words('english')]
        review = ' '.join(review)
    
        reviews.append(review)
    
    return reviews


# Call function   
reviews_1 = preprocessing(data)

type(reviews_1)

# Converting to dataframe and then csv file of pre processed data

reviews_pre = pd.DataFrame(reviews_1)

# Renaming Columns
reviews_pre.rename(columns={0:"Reviews"}, inplace=True)


# Export Pre-processed CSV

reviews_pre.to_csv('/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/Data/Pre-Processed.csv')

# Pre-Processed Data Preview

reviews_pre
