==> Sentimental analysis Project/5. Model Training (Logistic Regression, Naive Bayes, SVM) with TF-IDF.ipynb <==
# Mount Drive

from google.colab import drive
drive.mount('/content/drive')

<h3>Import Libraries and Data</h3>

# import libraries

import pandas as pd

# import pre-processed reviews and labels

# Reviews
reviews  = pd.read_csv('/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/Data/Pre-Processed.csv')
reviews.drop(['Unnamed: 0'],axis=1,inplace=True)

# Labels
labels = pd.read_csv('/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/Data/Amazon_Customer_Reviews.csv')

<h3> Label Coding</h3>

# dummyfying output variable

labels = pd.get_dummies(labels['Ratings'],drop_first=True)

<h3>Train-Test Split</h3>

# Train Test Split

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(reviews, labels, test_size = 0.20, random_state = 42)

# Appending each review to list, so that it becomes iterable and compatible for count vectorizer

X_train_list = []
X_test_list =[]

for index, row in X_train.iterrows():
  X_train_list.append(row['Reviews'])

for index, row in X_test.iterrows():
  X_test_list.append(row['Reviews'])

<h3>TF-IDF</h3>

# TFIDF using TFIDFVectorizer    

from sklearn.feature_extraction.text import TfidfVectorizer
tfidf = TfidfVectorizer(max_features=5000)

# Vectorizing train reviews (TF-IDF)

X_train_tfidf = tfidf.fit_transform(X_train_list).toarray()

# Vectorizing test reviews (TF-IDF)

X_test_tfidf = tfidf.fit_transform(X_test_list).toarray()

<h3>Naive Bayes Classifier</h3>

# Training model using Naive bayes classifier
from sklearn.naive_bayes import MultinomialNB
nb_model = MultinomialNB(alpha=0.01).fit(X_train_tfidf, y_train)

y_pred=nb_model.predict(X_test_tfidf)

# Training Evaluation

y_pred_train = nb_model.predict(X_train_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_train = confusion_matrix(y_pred_train, y_train)
print("Confusion Matrix:")
print(cm_train)

print("____________________________________________________________")

report_train = classification_report(y_pred_train, y_train)
print("Classification Report:",)
print (report_train)

print("____________________________________________________________")

accuracy_train = accuracy_score(y_pred_train,y_train)
print("Accuracy:",accuracy_train)

# Testing Evaluation Evaluation

y_pred_test = nb_model.predict(X_test_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_test = confusion_matrix(y_pred_test, y_test)
print("Confusion Matrix:")
print(cm_test)

print("____________________________________________________________")

report_test = classification_report(y_pred_test, y_test)
print("Classification Report:",)
print (report_test)

print("____________________________________________________________")

accuracy_test = accuracy_score(y_pred_test,y_test)
print("Accuracy:",accuracy_test)

<h3>Logistic Regression</h3>

# Training model using Logistic regression
from sklearn.linear_model import LogisticRegression
log_model = LogisticRegression().fit(X_train_tfidf, y_train)

y_pred=log_model.predict(X_test_tfidf)

# Training Evaluation

y_pred_train = log_model.predict(X_train_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_train = confusion_matrix(y_pred_train, y_train)
print("Confusion Matrix:")
print(cm_train)

print("____________________________________________________________")

report_train = classification_report(y_pred_train, y_train)
print("Classification Report:",)
print (report_train)

print("____________________________________________________________")

accuracy_train = accuracy_score(y_pred_train, y_train)
print("Accuracy:",accuracy_train)

# Testing Evaluation Evaluation

y_pred_test = log_model.predict(X_test_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_test = confusion_matrix(y_pred_test, y_test)
print("Confusion Matrix:")
print(cm_test)

print("____________________________________________________________")

report_test = classification_report(y_pred_test, y_test)
print("Classification Report:",)
print (report_test)

print("____________________________________________________________")

accuracy_test = accuracy_score(y_pred_test,y_test)
print("Accuracy:",accuracy_test)

<h3>Support Vector Machine

# Training model using Support Vector Machine

from sklearn import svm
svm_model = svm.SVC()
svm_model.fit(X_train_tfidf,y_train)

y_pred=svm_model.predict(X_test_tfidf)

# Training Evaluation

y_pred_train = svm_model.predict(X_train_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_train = confusion_matrix(y_pred_train, y_train)
print("Confusion Matrix:")
print(cm_train)

print("____________________________________________________________")

report_train = classification_report(y_pred_train, y_train)
print("Classification Report:",)
print (report_train)

print("____________________________________________________________")

accuracy_train = accuracy_score(y_pred_train,y_train)
print("Accuracy:",accuracy_train)

# Testing Evaluation Evaluation

y_pred_test = svm_model.predict(X_test_tfidf)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_test = confusion_matrix(y_pred_test, y_test)
print("Confusion Matrix:")
print(cm_test)

print("____________________________________________________________")

report_test = classification_report(y_pred_test, y_test)
print("Classification Report:",)
print (report_test)

print("____________________________________________________________")

accuracy_test = accuracy_score(y_pred_test,y_test)
print("Accuracy:",accuracy_test)

<h3>NGram TF-IDF</h3>


# ngram level tf-idf 
tfidf_vect_ngram = TfidfVectorizer(analyzer='word', token_pattern=r'\w{1,}', ngram_range=(2,3), max_features=5000)
X_train_tfidf_n = tfidf_vect_ngram.fit_transform(X_train_list).toarray()

X_test_tfidf_n = tfidf_vect_ngram.fit_transform(X_test_list).toarray()

# Training model using Naive bayes classifier
from sklearn.naive_bayes import MultinomialNB
nb_model = MultinomialNB(alpha=0.7).fit(X_train_tfidf_n, y_train)

y_pred=nb_model.predict(X_test_tfidf_n)

# Training Evaluation

y_pred_train = nb_model.predict(X_train_tfidf_n)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_train = confusion_matrix(y_pred_train, y_train)
print("Confusion Matrix:")
print(cm_train)


print("____________________________________________________________")

report_train = classification_report(y_pred_train, y_train)
print("Classification Report:",)
print (report_train)


print("____________________________________________________________")

accuracy_train = accuracy_score(y_pred_train,y_train)
print("Accuracy:",accuracy_train)

# Testing Evaluation Evaluation

y_pred_test = nb_model.predict(X_test_tfidf_n)

# Classification metrics
from sklearn.metrics import classification_report,confusion_matrix,accuracy_score
cm_test = confusion_matrix(y_pred_test, y_test)
print("Confusion Matrix:")
print(cm_test)

print("____________________________________________________________")

report_test = classification_report(y_pred_test, y_test)
print("Classification Report:",)
print (report_test)

print("____________________________________________________________")

accuracy_test = accuracy_score(y_pred_test,y_test)
print("Accuracy:",accuracy_test)

# Hyperparameter tuning

from sklearn.model_selection import RandomizedSearchCV

# Create the random grid
params = {'alpha': [0.01, 0.1, 0.5, 1.0, 10.0 ]}



# create classifier
nb_model= MultinomialNB()
nb_randomcv = RandomizedSearchCV(estimator=nb_model,param_distributions=params,n_iter=300,cv=10,verbose=2,
                               random_state=100,n_jobs=-1)

### fit the randomized model
nb_randomcv.fit(X_train_tfidf_n,y_train)

# Provides best parameters
nb_randomcv.best_params_


best_random_grid=nb_randomcv.best_estimator_
