==> Sentimental analysis Project/6. Model Training (LSTM and Bi-LSTM) with Word to Vec.ipynb <==
<h3>Import Libraries and Data

# Mount Drive

from google.colab import drive
drive.mount('/content/drive')

# Import libraries

from sklearn import model_selection, preprocessing, linear_model, naive_bayes, metrics, svm
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn import decomposition, ensemble

import pandas as pd
import xgboost, numpy, textblob, string
from keras.preprocessing import text, sequence
from keras import layers, models, optimizers

# Read Data

path = '/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/Data/Amazon_Customer_Reviews.csv'
data = pd.read_csv(path)
data.drop('Unnamed: 0',axis=1,inplace=True)
data.head()


<h3>Pre-Processing

#Data cleaning and preprocessing
import re
from nltk.stem import WordNetLemmatizer
wordnet = WordNetLemmatizer()

# Function for preprocessing
def preprocessing(data):    
    
    reviews =  []
    for i in range(0, len(data)):
        # Replacing values in rows
        review = data['Reviews'][i]
        review = review.replace('$','dollars')
        review = review.lower()
        reviews.append(review)
    
    return reviews
        
reviews_1 = preprocessing(data)

# Create Dataframe

df = pd.DataFrame(reviews_1)

# Rename Column

df.columns=['reviews']

df.head()

<h3> Word to vec Representation

# Import pretrained fasttext word to vec representation

path_vec = '/content/drive/MyDrive/Sentiment analysis/Natural-Language-Processing-main/Sentimental analysis Project/wiki-news-300d-1M.vec'

# load the pre-trained word-embedding vectors

embeddings_index = {}
for i, line in enumerate(open(path_vec)):
    values = line.split()
    embeddings_index[values[0]] = numpy.asarray(values[1:], dtype='float32')

embeddings_index

#dummyfying output variable
y=pd.get_dummies(data['Ratings'],drop_first=True)

<h3> Train Test Split

# split the dataset into training and validation datasets 
train_x, valid_x, train_y, valid_y = model_selection.train_test_split(df['reviews'], y)

<h3> Tokenizing Train Data


# create a tokenizer 
token = text.Tokenizer()
token.fit_on_texts(train_x)
word_index = token.word_index

# create token-embedding mapping
embedding_matrix = numpy.zeros((len(word_index) + 1, 300))
for word, i in word_index.items():
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector

embedding_matrix

<h3 >Tokenizing Test Data

# create a tokenizer 
token2 = text.Tokenizer()
token2.fit_on_texts(valid_x)
word_index2 = token2.word_index

# create token-embedding mapping
embedding_matrix2 = numpy.zeros((len(word_index2) + 1, 300))
for word, i in word_index2.items():
    embedding_vector2 = embeddings_index.get(word)
    if embedding_vector2 is not None:
        embedding_matrix2[i] = embedding_vector2

embedding_matrix2

<h3> Padding sequences

from tensorflow.keras.utils import pad_sequences

# convert text to sequence of tokens and pad them to ensure equal length vectors 
train_seq_x = pad_sequences(token.texts_to_sequences(train_x), maxlen=165)
valid_seq_x = pad_sequences(token2.texts_to_sequences(valid_x), maxlen=165)

train_seq_x.shape

valid_seq_x.shape

<h3> Evaluation Metrics

# Calculation of F1, Precision, recall

from keras import backend as K
def recall_m(y_true, y_pred):
    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
    recall = true_positives / (possible_positives + K.epsilon())
    return recall

def precision_m(y_true, y_pred):
    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
    predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
    precision = true_positives / (predicted_positives + K.epsilon())
    return precision

def f1_m(y_true, y_pred):
    precision = precision_m(y_true, y_pred)
    recall = recall_m(y_true, y_pred)
    return 2*((precision*recall)/(precision+recall+K.epsilon()))


<h3>LSTM

## Creating model
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding
from tensorflow.keras.layers import Dropout
from tensorflow.keras.layers import LSTM
from tensorflow.keras.layers import Dense

model=Sequential()
model.add(Embedding(len(word_index) + 1, 300, weights=[embedding_matrix],input_length=165 ,trainable=False))
model.add(Dropout(0.3))
model.add(LSTM(100))
model.add(Dense(1,activation='sigmoid'))
print(model.summary())



# compile the model
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc',f1_m,precision_m, recall_m])

# fit the model
history = model.fit(train_seq_x,train_y,validation_data=(valid_seq_x,valid_y),epochs=10,batch_size=64)


# evaluate the model
loss, accuracy, f1_score, precision, recall = model.evaluate(valid_seq_x,valid_y, verbose=0)

f1_score

<h3>Bi-Directional</h3>

from tensorflow.keras.layers import Bidirectional
## Creating model
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding
from tensorflow.keras.layers import Dropout
from tensorflow.keras.layers import GRU
from tensorflow.keras.layers import Dense

model=Sequential()
model.add(Embedding(len(word_index) + 1, 300, weights=[embedding_matrix],input_length=165 ,trainable=False))
model.add(Dropout(0.3))
model.add(Bidirectional(LSTM(100)))
model.add(Dense(1,activation='sigmoid'))
print(model.summary())



# compile the model
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc',f1_m,precision_m, recall_m])

# fit the model
history = model.fit(train_seq_x,train_y,validation_data=(valid_seq_x,valid_y),epochs=10,batch_size=64)



# evaluate the model
loss, accuracy, f1_score, precision, recall = model.evaluate(valid_seq_x,valid_y, verbose=0)

f1_score
